Subclass SymbolSpec, Production and Item from int
Instances of those classes are explicitly numbered and a huge number of
sorting and comparison takes place with calls to overloaded comparison
operators.  A better idea is to subclass from int and initialize with
counter value so that fast built-in comparison operators can be used.

Full generation time before changes:

  > time python3 testparse.py

  real    0m37.819s
  user    0m40.579s
  sys     0m0.166s

Full generation time after changes:

  > time python3 testparse.py

  real    0m21.821s
  user    0m21.320s
  sys     0m0.114s